move the area filter dropdowns
