+++ frontend/fingertips-frontend/components/organisms/AreaFilter/index.tsx
@@ -3,24 +3,12 @@ import {
   AreaType,
   AreaWithRelations,
 } from '@/generated-sources/ft-api-client';
-import {
-  SearchParams,
-  SearchStateManager,
-  SearchStateParams,
-} from '@/lib/searchStateManager';
-
-import {
-  Checkbox,
-  FormGroup,
-  H3,
-  LabelText,
-  SectionBreak,
-  Select,
-} from 'govuk-react';
-import { usePathname, useRouter } from 'next/navigation';
+import { SearchStateParams } from '@/lib/searchStateManager';
+import { H3, SectionBreak } from 'govuk-react';
 import styled from 'styled-components';
 import { ShowHideContainer } from '@/components/molecules/ShowHideContainer';
 import { SelectedAreasPanel } from '@/components/molecules/SelectedAreasPanel';
+import { SelectAreasFilterPanel } from '@/components/molecules/SelectAreasFilterPanel';
 
 interface AreaFilterProps {
   selectedAreasData?: AreaWithRelations[];
@@ -46,23 +34,6 @@ const StyledFilterDiv = styled('div')({
   padding: '1.5em 1em',
 });
 
-const StyledFilterLabel = styled(LabelText)({
-  fontWeight: 'bold',
-});
-
-const StyledFilterSelect = styled(Select)({
-  span: {
-    fontWeight: 'bold',
-  },
-  select: {
-    width: '100%',
-  },
-  marginBottom: '1em',
-});
-
-const isAreaSelected = (areaCode: string, selectedAreas?: Area[]): boolean =>
-  selectedAreas ? selectedAreas?.some((area) => area.code === areaCode) : false;
-
 export function AreaFilter({
   selectedAreasData,
   availableAreaTypes,
@@ -71,56 +42,6 @@ export function AreaFilter({
   availableAreas,
   searchState,
 }: Readonly<AreaFilterProps>) {
-  const pathname = usePathname();
-  const { replace } = useRouter();
-
-  const searchStateManager = SearchStateManager.initialise(searchState);
-
-  const areaTypeSelected = (valueSelected: string) => {
-    searchStateManager.addParamValueToState(
-      SearchParams.AreaTypeSelected,
-      valueSelected
-    );
-    searchStateManager.removeParamValueFromState(
-      SearchParams.GroupTypeSelected
-    );
-    searchStateManager.removeParamValueFromState(SearchParams.GroupSelected);
-    replace(searchStateManager.generatePath(pathname), { scroll: false });
-  };
-
-  const groupTypeSelected = (valueSelected: string) => {
-    searchStateManager.addParamValueToState(
-      SearchParams.GroupTypeSelected,
-      valueSelected
-    );
-    searchStateManager.removeParamValueFromState(SearchParams.GroupSelected);
-    replace(searchStateManager.generatePath(pathname), { scroll: false });
-  };
-
-  const groupSelected = (valueSelected: string) => {
-    searchStateManager.addParamValueToState(
-      SearchParams.GroupSelected,
-      valueSelected
-    );
-    replace(searchStateManager.generatePath(pathname), { scroll: false });
-  };
-
-  const handleAreaSelected = (areaCode: string, checked: boolean) => {
-    if (checked) {
-      searchStateManager.addParamValueToState(
-        SearchParams.AreasSelected,
-        areaCode
-      );
-    } else {
-      searchStateManager.removeParamValueFromState(
-        SearchParams.AreasSelected,
-        areaCode
-      );
-    }
-
-    replace(searchStateManager.generatePath(pathname), { scroll: false });
-  };
-
   return (
     <StyledFilterPane data-testid="area-filter-container">
       <StyledFilterPaneHeader>
@@ -134,77 +55,14 @@ export function AreaFilter({
         />
 
         <ShowHideContainer summary="Add or change areas">
-          <StyledFilterSelect
-            label="Select an area type"
-            data-testid="area-type-selector-container"
-            input={{
-              onChange: (e) => areaTypeSelected(e.target.value),
-              defaultValue: searchState?.[SearchParams.AreaTypeSelected],
-              disabled: selectedAreasData && selectedAreasData?.length > 0,
-            }}
-          >
-            {availableAreaTypes?.map((areaType) => (
-              <option key={areaType.key} value={areaType.key}>
-                {areaType.name}
-              </option>
-            ))}
-          </StyledFilterSelect>
-
-          <StyledFilterSelect
-            label="Select a group type"
-            data-testid="group-type-selector-container"
-            input={{
-              onChange: (e) => groupTypeSelected(e.target.value),
-              defaultValue: searchState?.[SearchParams.GroupTypeSelected],
-              disabled: selectedAreasData && selectedAreasData?.length > 0,
-            }}
-          >
-            {availableGroupTypes?.map((areaType) => (
-              <option key={areaType.key} value={areaType.key}>
-                {areaType.name}
-              </option>
-            ))}
-          </StyledFilterSelect>
-
-          <StyledFilterSelect
-            label="Select a group"
-            input={{
-              onChange: (e) => groupSelected(e.target.value),
-              defaultValue: searchState?.[SearchParams.GroupSelected],
-              disabled: selectedAreasData && selectedAreasData?.length > 0,
-            }}
-          >
-            {availableGroups?.map((area) => (
-              <option key={area.code} value={area.code}>
-                {area.name}
-              </option>
-            ))}
-          </StyledFilterSelect>
-
-          <FormGroup>
-            <StyledFilterLabel>Select one or more areas</StyledFilterLabel>
-            {availableAreas?.map((area) => {
-              const isAreaSelectedValue = isAreaSelected(
-                area.code,
-                selectedAreasData
-              );
-
-              return (
-                <Checkbox
-                  key={area.code}
-                  value={area.code}
-                  sizeVariant="SMALL"
-                  name="area"
-                  defaultChecked={isAreaSelectedValue}
-                  onChange={(e) =>
-                    handleAreaSelected(area.code, e.target.checked)
-                  }
-                >
-                  {area.name}
-                </Checkbox>
-              );
-            })}
-          </FormGroup>
+          <SelectAreasFilterPanel
+            selectedAreasData={selectedAreasData}
+            availableAreaTypes={availableAreaTypes}
+            availableGroupTypes={availableGroupTypes}
+            availableGroups={availableGroups}
+            availableAreas={availableAreas}
+            searchState={searchState}
+          />
         </ShowHideContainer>
       </StyledFilterDiv>
     </StyledFilterPane>
